+++ src/component/app/main/control/switch/LanguageSwitch.module.css
@@ -13,24 +13,7 @@
     vertical-align: middle;
 }
 
-.switch {
-    list-style-type: none;
-}
-
-.switch > li {
-    cursor: pointer;
-}
-
-.languageListElement {
-    display: inline;
-    padding-right: 0.4vw;
-    padding-left: 0.4vw;
-}
-
-.locales {
-}
-
-@media only screen and (max-width: 650px) {
+@media only screen and (max-width: 640px) {
     .localesContainer {
         color: var(--text-color-gray);
         bottom: 0;
@@ -38,11 +21,6 @@
         left: 0;
         margin-right: 0;
         width: 100%;
-    }
-
-    .locales {
-        padding: 0;
-        text-align: center;
     }
 }
