+++ src/component/app/main/control/switch/locales/Locales.test.js
